src/config.py
import os

from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# Model configurations with fallbacks to defaults
DEFAULT_MODEL = os.getenv('DEFAULT_MODEL', 'hf.co/bartowski/Qwen2.5-7B-Instruct-GGUF:Q5_K_M')
DEFAULT_EMBEDDER = os.getenv('DEFAULT_EMBEDDER', 'nomic-embed-text:latest')



# System messages
SYSTEM_PROMPT = """You are a helpful assistant explaining technical documentation.

You have access to the user's background to calibrate your explanations, but you should not reference or mention their background in your responses.

When responding:
- Adapt your technical language and complexity based on the background information provided
- Use only the provided documentation content to answer questions
- Do not reference or mention any personal information or background
- If information is not in the documentation, simply state that it's not covered in the available documentation

Focus on explaining the documentation content."""
