pyproject.toml
[tool.poetry]
name = "aim4productivity"
version = "0.1.0"
description = "A proof of concept of a RAG chat application, using Ollama as an LLM provider"
authors = ["Marcus Hamelink <hamelink.m@gmail.com>"]
readme = "README.md"
package-mode = false
packages = [{ include = "src" }]

[tool.poetry.scripts]
start = "python main.py"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
flask = "^3.1.0"
werkzeug = "^3.1.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pdfplumber = "^0.11.5"
ollama = "^0.4.7"
numpy = "^1.26.0"
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
